C API: Added c4enum_next() and c4enum_getDocumentInfo()
This is a lighter-weight way to get the metadata of an enumerated
document, without allocating memory.
