+++ C/c4Database.cc
@@ -373,136 +373,3 @@ bool c4raw_put(C4Database* database,
         database->endTransaction(false, NULL);
     return false;
 }
-
-
-#pragma mark - DOC ENUMERATION:
-
-CBFOREST_API const C4EnumeratorOptions kC4DefaultEnumeratorOptions = {
-    0, // skip
-    kC4InclusiveStart | kC4InclusiveEnd | kC4IncludeNonConflicted | kC4IncludeBodies
-};
-
-
-struct C4DocEnumerator {
-    C4Database *_database;
-    DocEnumerator _e;
-    C4EnumeratorOptions _options;
-
-    C4DocEnumerator(C4Database *database,
-                    sequence start,
-                    sequence end,
-                    const C4EnumeratorOptions &options)
-    :_database(database),
-     _e(*database, start, end, allDocOptions(options)),
-     _options(options)
-    { }
-
-    C4DocEnumerator(C4Database *database,
-                    C4Slice startDocID,
-                    C4Slice endDocID,
-                    const C4EnumeratorOptions &options)
-    :_database(database),
-     _e(*database, startDocID, endDocID, allDocOptions(options)),
-     _options(options)
-    { }
-
-    C4DocEnumerator(C4Database *database,
-                    std::vector<std::string>docIDs,
-                    const C4EnumeratorOptions &options)
-    :_database(database),
-     _e(*database, docIDs, allDocOptions(options)),
-     _options(options)
-    { }
-
-    static DocEnumerator::Options allDocOptions(const C4EnumeratorOptions &c4options) {
-        auto options = DocEnumerator::Options::kDefault;
-        options.skip = (unsigned)c4options.skip;
-        options.descending = (c4options.flags & kC4Descending) != 0;
-        options.inclusiveStart = (c4options.flags & kC4InclusiveStart) != 0;
-        options.inclusiveEnd = (c4options.flags & kC4InclusiveEnd) != 0;
-        if ((c4options.flags & kC4IncludeBodies) == 0)
-            options.contentOptions = KeyStore::kMetaOnly;
-        return options;
-    }
-    
-    C4Document* next() {
-        do {
-            if (!_e.next())
-                return NULL;
-        } while (!useDoc());
-        return newC4Document(_database, _e.doc());
-    }
-
-    inline bool useDoc() {
-        auto optFlags = _options.flags;
-        if ((optFlags & kC4IncludeDeleted) && (optFlags & kC4IncludeNonConflicted))
-            return true;
-        VersionedDocument::Flags docFlags;
-        revid revID;
-        slice docType;
-        if (!VersionedDocument::readMeta(_e.doc(), docFlags, revID, docType))
-            return false;
-        return (optFlags & kC4IncludeDeleted       || !(docFlags & VersionedDocument::kDeleted))
-            && (optFlags & kC4IncludeNonConflicted ||  (docFlags & VersionedDocument::kConflicted));
-    }
-};
-
-
-void c4enum_free(C4DocEnumerator *e) {
-    delete e;
-}
-
-
-C4DocEnumerator* c4db_enumerateChanges(C4Database *database,
-                                       C4SequenceNumber since,
-                                       const C4EnumeratorOptions *c4options,
-                                       C4Error *outError)
-{
-    try {
-        return new C4DocEnumerator(database, since+1, UINT64_MAX,
-                                   c4options ? *c4options : kC4DefaultEnumeratorOptions);
-    } catchError(outError);
-    return NULL;
-}
-
-
-C4DocEnumerator* c4db_enumerateAllDocs(C4Database *database,
-                                       C4Slice startDocID,
-                                       C4Slice endDocID,
-                                       const C4EnumeratorOptions *c4options,
-                                       C4Error *outError)
-{
-    try {
-        return new C4DocEnumerator(database, startDocID, endDocID,
-                                   c4options ? *c4options : kC4DefaultEnumeratorOptions);
-    } catchError(outError);
-    return NULL;
-}
-
-
-C4DocEnumerator* c4db_enumerateSomeDocs(C4Database *database,
-                                        C4Slice docIDs[],
-                                        size_t docIDsCount,
-                                        const C4EnumeratorOptions *c4options,
-                                        C4Error *outError)
-{
-    try {
-        std::vector<std::string> docIDStrings;
-        for (size_t i = 0; i < docIDsCount; ++i)
-            docIDStrings.push_back((std::string)docIDs[i]);
-        return new C4DocEnumerator(database, docIDStrings,
-                                   c4options ? *c4options : kC4DefaultEnumeratorOptions);
-    } catchError(outError);
-    return NULL;
-}
-
-
-C4Document* c4enum_nextDocument(C4DocEnumerator *e, C4Error *outError) {
-    try {
-        auto c4doc = e->next();
-        if (!c4doc)
-            recordError(FDB_RESULT_SUCCESS, outError);      // end of iteration is not an error
-        return c4doc;
-    } catchError(outError)
-    return NULL;
-}
