+++ C/c4DocEnumerator.cc
@@ -1,0 +1,195 @@
+//
+//  c4DocEnumerator.cc
+//  CBForest
+//
+//  Created by Jens Alfke on 12/16/15.
+//  Copyright © 2015 Couchbase. All rights reserved.
+//
+
+#include "c4Impl.hh"
+#include "c4DocEnumerator.h"
+
+#include "Database.hh"
+#include "Document.hh"
+#include "DocEnumerator.hh"
+#include "LogInternal.hh"
+#include "VersionedDocument.hh"
+#include <assert.h>
+
+using namespace forestdb;
+
+
+#pragma mark - DOC ENUMERATION:
+
+CBFOREST_API const C4EnumeratorOptions kC4DefaultEnumeratorOptions = {
+    0, // skip
+    kC4InclusiveStart | kC4InclusiveEnd | kC4IncludeNonConflicted | kC4IncludeBodies
+};
+
+
+struct C4DocEnumerator {
+    C4Database *_database;
+    DocEnumerator _e;
+    C4EnumeratorOptions _options;
+    C4DocumentFlags _docFlags;
+    revid _docRevID;
+    alloc_slice _docRevIDExpanded;
+
+    C4DocEnumerator(C4Database *database,
+                    sequence start,
+                    sequence end,
+                    const C4EnumeratorOptions &options)
+    :_database(database),
+     _e(*asDatabase(database), start, end, allDocOptions(options)),
+     _options(options)
+    { }
+
+    C4DocEnumerator(C4Database *database,
+                    C4Slice startDocID,
+                    C4Slice endDocID,
+                    const C4EnumeratorOptions &options)
+    :_database(database),
+     _e(*asDatabase(database), startDocID, endDocID, allDocOptions(options)),
+     _options(options)
+    { }
+
+    C4DocEnumerator(C4Database *database,
+                    std::vector<std::string>docIDs,
+                    const C4EnumeratorOptions &options)
+    :_database(database),
+     _e(*asDatabase(database), docIDs, allDocOptions(options)),
+     _options(options)
+    { }
+
+    static DocEnumerator::Options allDocOptions(const C4EnumeratorOptions &c4options) {
+        auto options = DocEnumerator::Options::kDefault;
+        options.skip = (unsigned)c4options.skip;
+        options.descending = (c4options.flags & kC4Descending) != 0;
+        options.inclusiveStart = (c4options.flags & kC4InclusiveStart) != 0;
+        options.inclusiveEnd = (c4options.flags & kC4InclusiveEnd) != 0;
+        if ((c4options.flags & kC4IncludeBodies) == 0)
+            options.contentOptions = KeyStore::kMetaOnly;
+        return options;
+    }
+
+    bool next() {
+        do {
+            if (!_e.next())
+                return false;
+        } while (!useDoc());
+        return true;
+    }
+
+    C4Document* getDoc() {
+        return _e ? newC4Document(_database, _e.doc()) : NULL;
+    }
+
+    bool getDocInfo(C4DocumentInfo *outInfo) {
+        if (!_e)
+            return false;
+        outInfo->docID = _e.doc().key();
+        _docRevIDExpanded = _docRevID.expanded();
+        outInfo->revID = _docRevIDExpanded;
+        outInfo->flags = _docFlags;
+        outInfo->sequence = _e.doc().sequence();
+        return true;
+    }
+
+private:
+    inline bool useDoc() {
+        slice docType;
+        if (!_e.doc().exists()) {
+            // Client must be enumerating a list of docIDs, and this doc doesn't exist.
+            // Return it anyway, without the kExists flag.
+            _docFlags = 0;
+            _docRevID = revid();
+            return true;
+        }
+        VersionedDocument::Flags flags;
+        if (!VersionedDocument::readMeta(_e.doc(), flags, _docRevID, docType))
+            return false;
+        _docFlags = (C4DocumentFlags)flags | kExists;
+        auto optFlags = _options.flags;
+        return (optFlags & kC4IncludeDeleted       || !(_docFlags & VersionedDocument::kDeleted))
+            && (optFlags & kC4IncludeNonConflicted ||  (_docFlags & VersionedDocument::kConflicted));
+    }
+};
+
+
+void c4enum_free(C4DocEnumerator *e) {
+    delete e;
+}
+
+
+C4DocEnumerator* c4db_enumerateChanges(C4Database *database,
+                                       C4SequenceNumber since,
+                                       const C4EnumeratorOptions *c4options,
+                                       C4Error *outError)
+{
+    try {
+        return new C4DocEnumerator(database, since+1, UINT64_MAX,
+                                   c4options ? *c4options : kC4DefaultEnumeratorOptions);
+    } catchError(outError);
+    return NULL;
+}
+
+
+C4DocEnumerator* c4db_enumerateAllDocs(C4Database *database,
+                                       C4Slice startDocID,
+                                       C4Slice endDocID,
+                                       const C4EnumeratorOptions *c4options,
+                                       C4Error *outError)
+{
+    try {
+        return new C4DocEnumerator(database, startDocID, endDocID,
+                                   c4options ? *c4options : kC4DefaultEnumeratorOptions);
+    } catchError(outError);
+    return NULL;
+}
+
+
+C4DocEnumerator* c4db_enumerateSomeDocs(C4Database *database,
+                                        const C4Slice docIDs[],
+                                        size_t docIDsCount,
+                                        const C4EnumeratorOptions *c4options,
+                                        C4Error *outError)
+{
+    try {
+        std::vector<std::string> docIDStrings;
+        for (size_t i = 0; i < docIDsCount; ++i)
+            docIDStrings.push_back((std::string)docIDs[i]);
+        return new C4DocEnumerator(database, docIDStrings,
+                                   c4options ? *c4options : kC4DefaultEnumeratorOptions);
+    } catchError(outError);
+    return NULL;
+}
+
+
+bool c4enum_next(C4DocEnumerator *e, C4Error *outError) {
+    try {
+        if (e->next())
+            return true;
+        recordError(FDB_RESULT_SUCCESS, outError);      // end of iteration is not an error
+    } catchError(outError)
+    return false;
+}
+
+
+bool c4enum_getDocumentInfo(C4DocEnumerator *e, C4DocumentInfo *outInfo) {
+    return e->getDocInfo(outInfo);
+}
+
+
+C4Document* c4enum_getDocument(C4DocEnumerator *e, C4Error *outError) {
+    try {
+        auto c4doc = e->getDoc();
+        if (!c4doc)
+            recordError(FDB_RESULT_SUCCESS, outError);      // end of iteration is not an error
+        return c4doc;
+    } catchError(outError)
+    return NULL;
+}
+
+C4Document* c4enum_nextDocument(C4DocEnumerator *e, C4Error *outError) {
+    return c4enum_next(e, outError) ? c4enum_getDocument(e, outError) : NULL;
+}
